requirements.txt
fastapi
uvicorn[standard]
httpx
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-multipart
pytest
pytest-asyncio
httpx
alembic
